requirements.txt
# Core LangGraph and LangChain
langgraph==0.2.28
langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

# OpenAI for LLM
openai==1.51.0

# API clients
requests==2.32.3
httpx==0.27.0

# Data handling
pydantic==2.9.0
python-dotenv==1.0.1
jsonschema==4.23.0

# Email
sendgrid==6.11.0

# Google Sheets
gspread==6.1.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1

# Vector storage (optional for memory)
chromadb==0.5.5

# Utilities
python-dateutil==2.9.0
pyyaml==6.0.2

# Development
pytest==8.3.3
black==24.8.0
